refactor!: migrate the code to `econnect_metronet` naming
