+++ .pre-commit-config.yaml
@@ -10,7 +10,7 @@ repos:
     - id: check-toml
     - id: check-yaml
     - id: end-of-file-fixer
-      exclude: custom_components/elmo_iess_alarm/manifest.json
+      exclude: custom_components/econnect_metronet/manifest.json
     - id: mixed-line-ending
     - id: trailing-whitespace
