+++ custom_components/elmo_iess_alarm/const.py
@@ -12,7 +12,7 @@ CONF_AREAS_ARM_HOME = "areas_arm_home"
 CONF_AREAS_ARM_NIGHT = "areas_arm_night"
 CONF_AREAS_ARM_VACATION = "areas_arm_vacation"
 CONF_SCAN_INTERVAL = "scan_interval"
-DOMAIN = "elmo_iess_alarm"
+DOMAIN = "econnect_metronet"
 KEY_DEVICE = "device"
 KEY_COORDINATOR = "coordinator"
 KEY_UNSUBSCRIBER = "options_unsubscriber"
