+++ custom_components/elmo_iess_alarm/exceptions.py
@@ -1,4 +1,3 @@
-"""Exceptions raised by elmo_iess_alarm component."""
 from homeassistant.exceptions import HomeAssistantError
